requirements.txt
flask
deepface
opencv-python
python-dotenv
tf-keras
numpy
pillow
requests
